+++ changelog.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `architect` Update @angular/core@13 @angular/cli@13.
 - `architect` Update @angular/cli@13 update @angular/material@13.
 - `feature` Add story service, but only show events.
+- `feature` Events and SchedulesItems are shown depending on ShowModes.
 
 ## [0.2.3] - Changes after review with partners